requirements.txt
numpy==1.18.1
pandas==0.25.3
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
xlrd==1.2.0
